requirements.txt
streamlit>=1.24.0
crewai_tools>=0.10.0
python-dotenv>=0.19.0
langchain>=0.0.329
openai>=1.3.0
selenium>=4.11.2
unstructured>=0.10.30
chromadb>=0.4.15
